home: fix responsiveness of sermons section
