+++ package.json
@@ -7,7 +7,6 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "lodash": "^4.17.21",
     "vue": "^3.0.5"
   },
   "devDependencies": {
